fix(代码生成): 高版本MySQL数据库最大长度超出整型范围导致无法保存。

+++ seezoon-generator/src/main/java/com/seezoon/generator/plan/ColumnPlan.java
@@ -54,7 +54,7 @@ public class ColumnPlan implements Comparable<ColumnPlan> {
     /**
      * 字段长度eg:64
      */
-    private Integer maxLength;
+    private Long maxLength;
 
     // 如下字段接收前端参数设置
     /**
